Cold wave duration index wrt mean of reference period

Let infile1 be a time series of the daily minimum temperature TN, and let infile2 be the mean TNnorm of daily minimum temperatures for any period used as reference. Then counted is the number of days where, in intervals of at least nday consecutive days, TN < TNnorm - T. The numbers nday and T are optional parameters with default nday = 6 and T = 5°C. A further output variable is the number of cold waves longer than or equal to nday days. TNnorm is calculated as the mean of minimum temperatures of a five day window centred on each calendar day of a given climate reference period. Note that both TN and TNnorm have to be given in the same units. The date information of a timestep in outfile is the date of the last contributing timestep in infile1.

Usage

cdo_eca_cwdi(ifile1, ifile2, nday = NULL, T = NULL, ofile = NULL)

Arguments

ifile1, ifile2

Strings with the path to the input files.

nday

INTEGER - Number of consecutive days (default: nday = 6)

T

FLOAT - Temperature offset (unit: °C; default: T = 5°C)

ofile

String with the path to the output file.