Monthly arithmetic

This module performs simple arithmetic of a time series and one timestep with the same month and year. For each field in infile1 the corresponding field of the timestep in infile2 with the same month and year is used. The input files need to have the same structure with the same variables. Usually infile2 is generated by an operator of the module MONSTAT.

Usage

cdo_monadd(ifile1, ifile2, ofile = NULL)

cdo_mondiv(ifile1, ifile2, ofile = NULL)

cdo_monmul(ifile1, ifile2, ofile = NULL)

cdo_monsub(ifile1, ifile2, ofile = NULL)

Arguments

ifile1, ifile2

Strings with the path to the input files.

ofile

String with the path to the output file.

Details

monadd  Add monthly time series
        Adds a time series and a monthly time series.
monsub  Subtract monthly time series
        Subtracts a time series and a monthly time series.
monmul  Multiply monthly time series
        Multiplies a time series and a monthly time series.
mondiv  Divide monthly time series
        Divides a time series and a monthly time series.