Set missing value

This module sets part of a field to missing value or missing values to a constant value. Which part of the field is set depends on the chosen operator.

Usage

cdo_setctomiss(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setmisstoc(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setmisstodis(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setmisstonn(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setmissval(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setrtomiss(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

cdo_setvrange(
  ifile,
  neighbors = NULL,
  newmiss = NULL,
  c = NULL,
  rmin = NULL,
  rmax = NULL,
  ofile = NULL
)

Arguments

ifile

String with the path to the input file.

neighbors

INTEGER - Number of nearest neighbors

newmiss

FLOAT - New missing value

c

FLOAT - Constant

rmin

FLOAT - Lower bound

rmax

FLOAT - Upper bound

ofile

String with the path to the output file.

Details

setmissval    Set a new missing value
                       / newmiss   if i(t,x) EQ miss
              o(t,x) =
                       \\ i(t,x)    if i(t,x) NE miss
setctomiss    Set constant to missing value
                       / miss   if i(t,x) EQ c
              o(t,x) =
                       \\ i(t,x) if i(t,x) NE c
setmisstoc    Set missing value to constant
                       / c      if i(t,x) EQ miss
              o(t,x) =
                       \\ i(t,x) if i(t,x) NE miss
setrtomiss    Set range to missing value
                       / miss   if i(t,x) GE rmin AND i(t,x) LE rmax
              o(t,x) =
                       \\ i(t,x) if i(t,x) LT rmin OR  i(t,x) GT rmax
setvrange     Set valid range
                       / miss   if i(t,x) LT rmin OR  i(t,x) GT rmax
              o(t,x) =
                       \\ i(t,x) if i(t,x) GE rmin AND i(t,x) LE rmax
setmisstonn   Set missing value to nearest neighbor
              Set all missing values to the nearest non missing value.
                       / i(t,y) if i(t,x) EQ miss AND i(t,y) NE miss
              o(t,x) =
                       \\ i(t,x) if i(t,x) NE miss
setmisstodis  Set missing value to distance-weighted average
              Set all missing values to the distance-weighted average of the nearest non missing values.
              The default number of nearest neighbors is 4.